Collapse the project into `streams`
Since this is now open, it’s easier to manage this way.

- Combine Makefiles
- Update READMEs
- De-prefix environment variables
