+++ Makefile
@@ -1,26 +1,9 @@
 include Makefile.variable
 
 announce:
-	$(call becho,"=== Ello Go Project ===")
+	$(call becho,"=== Ello Streams Project ===")
 
-setup: announce get-tools
-	@$(PRINT_LINE)
-	$(call becho,"~~~    github.com/ello/ello-go/streams   ~~~")
-	@$(PRINT_LINE)
-	@cd streams && $(MAKE) setup
-
-get-tools: get-tools-ci
-	@brew rm --force fswatch readline > /dev/null 2>&1
-	@brew install fswatch readline > /dev/null 2>&1
-	@$(PRINT_OK)
-
-setup-ci: announce get-tools-ci
-	@$(PRINT_LINE)
-	$(call becho,"~~~    github.com/ello/ello-go/streams   ~~~")
-	@$(PRINT_LINE)
-	@cd streams && $(MAKE) setup
-
-get-tools-ci:
+get-tools:
 	@go get -u "github.com/Masterminds/glide"
 	@go build "github.com/Masterminds/glide"
 	@go get -u "github.com/alecthomas/gometalinter"
@@ -37,66 +20,71 @@ get-tools-ci:
 	@go get -u -f "github.com/alecthomas/gocyclo"
 	@go get -u -f "github.com/mvdan/interfacer/cmd/interfacer"
 	@go get -u -f "github.com/golang/lint/golint"
-	@$(PRINT_OK")
-
-
-# db-setup:
-# 	@echo "=== setup db ==="
-# 	@dropdb --if-exists ello_services_test > /dev/null 2>&1
-# 	@/bin/echo -n "."
-# 	@createdb ello_services_test
-# 	@/bin/echo -n "."
-# 	@psql ello_services_test < schema.dump > /dev/null 2>&1
-# 	@echo "√"
-
-# dependencies:export GO15VENDOREXPERIMENT=1
-# dependencies: clean
-# 	@echo "=== deps ==="
-# 	@glide install
-# 	@/bin/echo -n "√"
-
-# clean:
-# 	@echo "=== cleaning ==="
-# 	@rm -rf vendor
-# 	@/bin/echo -n "."
-# 	@rm -rf bin
-# 	@echo "√"
-
-# errcheck:
-# 	@echo "=== errcheck ==="
-# 	@errcheck github.com/ello/services/stream/...
-
+	@$(PRINT_OK)
 
 install:export GO15VENDOREXPERIMENT=1
 install: test
 	@echo "=== go install ==="
 	@go install -ldflags=$(GOLDFLAGS)
+	@$(PRINT_OK)
 
-all:
-	@$(PRINT_LINE)
-	$(call becho,"~~~    github.com/ello/ello-go/streams   ~~~")
-	@$(PRINT_LINE)
-	@cd streams && $(MAKE) all
+# From streams
 
-test:
-	@$(PRINT_LINE)
-	$(call becho,"~~~    github.com/ello/ello-go/streams   ~~~")
-	@$(PRINT_LINE)
-	@cd streams && $(MAKE) test
+all: test build docker
 
-test-w:
-	@echo "=== testing | watch mode ==="
-	@fswatch -o . -r | xargs -n1 -I{} make test
+setup: announce get-tools dependencies
+
+#TODO Try getting rid of the vendor flag env var after 1.6 is out
+dependencies:export GO15VENDOREXPERIMENT=1
+dependencies:
+	@glide install
+	@glide rebuild
+	@$(PRINT_OK)
+
+clean:
+	@rm -rf vendor
+	@rm -rf bin
+	@$(PRINT_OK)
+
+vet:export GO15VENDOREXPERIMENT=1
+vet:
+	@go vet `glide novendor`
+	@$(PRINT_OK)
+
+# TODO Re-enable these linters once vendor support is better (potentially 1.6)
+lint:export GO15VENDOREXPERIMENT=1
+lint:
+	@gometalinter --vendor  --deadline=10s --disable=gotype --disable=varcheck --disable=aligncheck --disable=structcheck --dupl-threshold=70 `glide novendor`
+	@$(PRINT_OK)
+
+fmt:export GO15VENDOREXPERIMENT=1
+fmt:
+	@gofmt -s -w `glide nv | sed 's/\.\.\./*.go/g' | sed 's/.\///'`
+	@$(PRINT_OK)
+
+build:export GO15VENDOREXPERIMENT=1
+build:
+	@mkdir -p bin/
+	@go build -ldflags $(GOLDFLAGS) -o bin/streams
+	@$(PRINT_OK)
+
+rebuild: clean build
+
+docker: test
+	@docker build -t streams . > /dev/null 2>&1
+	@$(PRINT_OK)
+
+test:export GO15VENDOREXPERIMENT=1
+test:export LOGXI=dat:sqlx=off
+test: fmt vet lint
+	@go test `glide novendor` -cover
+	@$(PRINT_OK)
 
 server:export GO15VENDOREXPERIMENT=1
 server:
-	@echo "=== server ==="
 	@go run main.go
 
 server-w:
-	@echo "=== server | watch mode ==="
 	@gin -a 8080 -i run
-
-deploy: test
 
 .PHONY: setup cloc errcheck vet lint fmt install build test deploy docker
