+++ model/stream.go
