+++ util/env_test.go
@@ -5,7 +5,7 @@ import (
 	"os"
 	"testing"
 
-	"github.com/ello/ello-go/streams/util"
+	"github.com/ello/streams/util"
 )
 
 func ExampleGetEnvWithDefault() {
